Major update to documentation of the new Plone 5 Resource Registry.
Atomic commit messages to follow:

update the introduction to improve the structure, clarify a few explanations, add links to external definitions of terms, and provide a bit more background for the concepts to follow

small improvement to wording

improve the clarity of the description of bundle registration.  Be more explicit about what is allowed, required, etc

improve the documentation about various options for resource registry records, include cross-links for better clarity

provide more numerous examples of different forms of resource records.  Include more textual explanation of the important attributes of each example

add a note about the difference between values for the js option and those for the css option

minor grammatical fixes

minor improvements to docs regarding resources

begin re-organizing bundle section of the documentation.  Make formatting and language around defining and configuring bundles consistent with the language used for resources.  Make section sub-headings more consistent

fix method signature to make it clear this is for resources, not bundles

improve the section on compiling bundles, including specific instructions on making the compiling script available.  Fix up the organization a bit to better match the resources section

improve documentation of the new plone traversal namespace

improve explanation of using your own build system for compiled bundle resources

improve structure and English syntax of the description of the mismatched anonymous define() error

fix errors and omissions noted by @seanupton

improvement to naming.  The 'static' part should not be included in the title, as it is specific to Plone's own registration of a folder for that namespace.

Update the section on Diazo Theme Bundles
Provide improved Bundle registration examples (more are needed)
Improve documentation of the resource request APIs
Fix up the build to migrating older code, including links to more complete documentation of module definition patterns.
Improve the syntax of the English in the section on including non-RequireJS resources in Diazo themes.

add a few more example bundles thanks to @ebrehault.  Also improve the descriptions of each to point out some differences in usage and the reasons for them

fix typo, thx @svx

fix external linking style according to styleguide.  Thx @svx for the pointer.

improve the link text a bit for this link in accordance with @svx's wishes.

follow the lead of @svx in #522 and spell it 'Javascript'

provide an example of registering a simple non-AMD javascript as a 'pre-compiled' bundle (thanks @ebrehault)

Revert "follow the lead of @svx in #522 and spell it 'Javascript'"

I misread the title of that commit and did exactly the wrong thing.

This reverts commit 09f9e586674d3ab4c5b3a049d8171f0a73fa8f24.

fix wording for clarity, thx @jensens

document the stub_js_resources option and mention it's use in the documentation for minimizing bundles.  Replace the docs for adding your stuff to the plone bundle as that is not recommended