Move view/update so messages will be seen
For the example code shown, "view/update" was moved to the top of the page template code so that informational messages generated by that code would be generated before the page is rendered. Prior to this, the "view/update" was located in the form declaration, and informational messages would not be seen immediately (would be seen if the page was refreshed, for example).