Fix implicit link to subsection.
In reStructuredText it is easily possible to create a link to a
subsection by adding backticks around the same words as the title of
the subsection.

Here, the question mark at the end was missing, resulting in a broken
link.

This has been fixed by this commit.

modified:   manage/upgrading/version_specific_migration/upgrade_zodb_to_python3.rst