Add backend installation documentation.
This is an edited copy of what I included in the recent release notes.  See:
https://dist.plone.org/release/6.0.0a4/RELEASE-NOTES.md

For the moment it includes a section on how to install Volto.
This is a candidate for removal because the Volto docs are included automatically and I point to it.
But the Volto docs also talk about how to install a backend.
So maybe keep for now and improve later.

I called the main file `installation.md`.
I am fine with renaming this to `installing.md` or `install.md` if that sounds more natural to a native speaker.
I will want to point to this file from the new release notes, so best pick a name now and stick with it.