Documentation for Discussion migration to 6.1.
This is for the PLIP that I have just merged:
https://github.com/plone/Products.CMFPlone/issues/3958

I changed the existing section title from "Discussion is disabled by default" to "Discussion is a core add-on".
Discussion was always disabled by default, so this would not be a change.

Note that the Discussion control panel is still missing from Volto: https://github.com/plone/volto/issues/29
But the upgrade to 6.1 does not change anything to this.