Don't mention installing add-ons here, activating is enough.
Sure, the Python package of an add-on must be installed before you can activate it, but the same is true for a distribution.
The plone.classicui distribution has basically the same Python package dependencies as CMFPlone, although it activates a few extra ones automatically if they are available (say plone.app.caching).